chore: Move lint after unit tests.
It is disruptive to code-build-test cycles to run lint on code that
may still be generating red tests. This is especially annoying when
you're adding or removing arguments to functions and the linter is
bitching at you about linefeeds when you don't even know if the
code works yet.

Should block PRs, not unit test runs.