feat: Narrow the bar display by another couple of characters by using ANSI line drawing.
test 1                                        ▏█████████████████████████▕ 125,555,696 ops/sec | 11 samples

By using a left and right justified equivalent of "|" there's enough space
to distinguish the bar from the end braces. These still look reasonable
pasted into Github issue comments (which was the problem with one of
my earlier attempts to make the bar display narrower).

Since the important info in the chart output is at the right end of
the line, this is what will probably be the last in my edits to
try to shift the punchline of the output as far left as is possible.