feat: Put the 'lifecycle' into lifecycle.js
index.js is doing a bunch of 'feature envy' on lifecycle.js and one
of the consequences of that is that useWorkers doesn't necessarily
get proper warmup runs in the worker thread.

By pushing the logic into lifecycle::runBenchmarks, we can do the
warming there. Special case for useWorkers=true getting a single
benchmark in the receiving worker thread instead of the entire array.

This should fix #136