bugfix: Fix warming and configuration change issues with detectDeadCodeElimination=true
Avoid configuration changes mid-run, and move warnings to the top of the
output.

As previously implemented, the worker mode warnings would be issued
once per benchmark, interlaced with them, and with DCE enabled wouldn't
issue a warning at all.

Putting the warnings in the middle of the test output is more likely
to be missed by a user, lost in the noise. This change also settles the
issue of which plugins we are running to the constructor, also avoiding

1) re-writing the benchmark string after creation
2) rewriting the benchmark string after warming has already supposedly completed